Add gen-cert script to generate certificates
Add a script to (re-)generate all certificates necessary for ratsd core,
and change the default protocol from HTTP to HTTPS

Ratsd certs are signed with a cert named "rootCA.crt" in the same
directory as the script (and the certs that will be generated). The
script will generate a self-signed rootCA.crt if one isn't provided.

Signed-off-by: Ian Chin Wang <ian.chin.wang@oracle.com>