Enable attester selections from the request
A new option is now added to config.yaml, list-options. If it's set to
"all", ratsd populates the EAT with CMW from all available attesters. If
it's set to "selected", the user must specify the name of each attester
along with the associated options in `attester-selection`. If the user
does not wish to specify the attester-specific option, "$attester_name":
"null" should be specified. The following is an example of the request
body.

```
{
  "nonce": "base64urlencoded",

  "attester-selection": {
    "attester-id-1": {
      "param11name": "param11value",
      "param12name": "param12value"
    },
    "attester-id-2": {
      "param21name": "param21value"
    },
    "attester-id-3": null
  }
}
```

Signed-off-by: Ian Chin Wang <ian.chin.wang@oracle.com>