fix(go): revert to older golang version (#49)
* fix(go): revert to older golang version

go 1.24 isn't available yet in major Fedora-like OSes. As such, we need to
revert golang to an older version to ensure the build could proceed with
packages the OS distributes, instead of downloading the latest from
external sources (needed for reproducible builds).

go 1.24 added support for "tool" directive in go.mod for tools needed
during the build phase. But this isn't available in older golang versions.
So, we need to use the tools.go pattern to ensure we download the build tools.

Adds an "install-tools" Makefile target to install build tools, and
updates README

Signed-off-by: Jagannathan Raman <jag.raman@oracle.com>